+++ include/sdf/Box.hh
@@ -20,9 +20,13 @@
 #include <ignition/math/Vector3.hh>
 #include <sdf/Error.hh>
 #include <sdf/Element.hh>
+#include <sdf/sdf_config.h>
 
 namespace sdf
 {
+  // Inline bracke to help doxygen filtering.
+  inline namespace SDF_VERSION_NAMESPACE {
+  //
   // Forward declare private data class.
   class BoxPrivate;
 
@@ -33,8 +37,21 @@ namespace sdf
     /// \brief Constructor
     public: Box();
 
+    /// \brief Copy constructor
+    /// \param[in] _box Box to copy.
+    public: Box(const Box &_box);
+
+    /// \brief Move constructor
+    /// \param[in] _box Box to move.
+    public: Box(Box &&_box);
+
     /// \brief Destructor
     public: virtual ~Box();
+
+    /// \brief Assignment operator.
+    /// \param[in] _box The box to set values from.
+    /// \return *this
+    public: Box &operator=(const Box &_box);
 
     /// \brief Load the box geometry based on a element pointer.
     /// This is *not* the usual entry point. Typical usage of the SDF DOM is
@@ -61,6 +78,7 @@ namespace sdf
     /// \brief Private data pointer.
     private: BoxPrivate *dataPtr;
   };
+  }
 }
 
 #endif