+++ include/sdf/Cylinder.hh
@@ -19,9 +19,14 @@
 
 #include <sdf/Error.hh>
 #include <sdf/Element.hh>
+#include <sdf/sdf_config.h>
 
 namespace sdf
 {
+  // Inline bracke to help doxygen filtering.
+  inline namespace SDF_VERSION_NAMESPACE {
+  //
+
   // Forward declare private data class.
   class CylinderPrivate;
 
@@ -32,8 +37,21 @@ namespace sdf
     /// \brief Constructor
     public: Cylinder();
 
+    /// \brief Copy constructor
+    /// \param[in] _cylinder Cylinder to copy.
+    public: Cylinder(const Cylinder &_cylinder);
+
+    /// \brief Move constructor
+    /// \param[in] _cylinder Cylinder to move.
+    public: Cylinder(Cylinder &&_cylinder);
+
     /// \brief Destructor
     public: virtual ~Cylinder();
+
+    /// \brief Assignment operator.
+    /// \param[in] _cylinder The cylinder to set values from.
+    /// \return *this
+    public: Cylinder &operator=(const Cylinder &_cylinder);
 
     /// \brief Load the cylinder geometry based on a element pointer.
     /// This is *not* the usual entry point. Typical usage of the SDF DOM is
@@ -68,5 +86,6 @@ namespace sdf
     /// \brief Private data pointer.
     private: CylinderPrivate *dataPtr;
   };
+  }
 }
 #endif