+++ include/sdf/Joint.hh
@@ -21,10 +21,15 @@
 #include <ignition/math/Pose3.hh>
 #include "sdf/Element.hh"
 #include "sdf/Types.hh"
+#include "sdf/sdf_config.h"
 #include "sdf/system_util.hh"
 
 namespace sdf
 {
+  // Inline bracke to help doxygen filtering.
+  inline namespace SDF_VERSION_NAMESPACE {
+  //
+
   // Forward declarations.
   class JointAxis;
   class JointPrivate;
@@ -137,7 +142,9 @@ namespace sdf
 
     /// \brief Get the pose of the joint. This is the pose of the joint
     /// as specified in SDF (<joint> <pose> ... </pose></joint>).
-    /// \return The pose of the joint.
+    /// Transformations have not been applied to the return value.
+    /// \return The pose of the joint. This is the raw pose value, as set in
+    /// the SDF file.
     public: const ignition::math::Pose3d &Pose() const;
 
     /// \brief Set the pose of the joint.
@@ -147,13 +154,13 @@ namespace sdf
 
     /// \brief Get the name of the coordinate frame in which this joint's
     /// pose is expressed. A empty value indicates that the frame is the
-    /// parent model.
+    /// child link frame.
     /// \return The name of the pose frame.
     public: const std::string &PoseFrame() const;
 
     /// \brief Set the name of the coordinate frame in which this joint's
     /// pose is expressed. A empty value indicates that the frame is the
-    /// parent model.
+    /// child link frame.
     /// \param[in] _frame The name of the pose frame.
     public: void SetPoseFrame(const std::string &_frame);
 
@@ -166,5 +173,6 @@ namespace sdf
     /// \brief Private data pointer.
     private: JointPrivate *dataPtr = nullptr;
   };
+  }
 }
 #endif