+++ include/sdf/Param.hh
@@ -33,6 +33,7 @@
 #include <ignition/math.hh>
 
 #include "sdf/Console.hh"
+#include "sdf/sdf_config.h"
 #include "sdf/system_util.hh"
 #include "sdf/Types.hh"
 
@@ -45,6 +46,10 @@
 
 namespace sdf
 {
+  // Inline bracket to help doxygen filtering.
+  inline namespace SDF_VERSION_NAMESPACE {
+  //
+
   class SDFORMAT_VISIBLE Param;
 
   /// \def ParamPtr
@@ -362,6 +367,7 @@ namespace sdf
   bool Param::IsType() const
   {
     return std::holds_alternative<Type>(this->dataPtr->value);
+  }
   }
 }