+++ include/sdf/Root.hh
@@ -21,10 +21,15 @@
 
 #include "sdf/SDFImpl.hh"
 #include "sdf/Types.hh"
+#include "sdf/sdf_config.h"
 #include "sdf/system_util.hh"
 
 namespace sdf
 {
+  // Inline bracket to help doxygen filtering.
+  inline namespace SDF_VERSION_NAMESPACE {
+  //
+
   // Forward declarations.
   class Model;
   class RootPrivate;
@@ -32,6 +37,10 @@ namespace sdf
 
   /// \brief Root class that acts as an entry point to the SDF document
   /// model.
+  ///
+  /// Multiple worlds can exist in a single SDF file. A user of multiple
+  /// worlds could run parallel instances of simulation, or offer selection
+  /// of a world at runtime.
   ///
   /// # Usage
   ///
@@ -121,5 +130,6 @@ namespace sdf
     /// \brief Private data pointer
     private: RootPrivate *dataPtr = nullptr;
   };
+  }
 }
 #endif