+++ src/parser.cc
@@ -37,6 +37,7 @@
 
 namespace sdf
 {
+inline namespace SDF_VERSION_NAMESPACE {
 // Class to handle Ruby initialization.
 class RubyInitializer
 {
@@ -396,7 +397,9 @@ bool readFile(const std::string &_filename, SDFPtr _sdf, Errors &_errors)
   }
 
   if (readDoc(&xmlDoc, _sdf, filename, true, _errors))
+  {
     return true;
+  }
   else if (sdf::URDF2SDF::IsURDF(filename))
   {
     sdf::URDF2SDF u2g;
@@ -848,7 +851,7 @@ bool readXml(TiXmlElement *_xml, ElementPtr _sdf, Errors &_errors)
 
   if (_sdf->GetCopyChildren())
   {
-    copyChildren(_sdf, _xml);
+    copyChildren(_sdf, _xml, false);
   }
   else
   {
@@ -1016,14 +1019,16 @@ bool readXml(TiXmlElement *_xml, ElementPtr _sdf, Errors &_errors)
 
       if (descCounter == _sdf->GetElementDescriptionCount())
       {
-        sdfwarn << "XML Element[" << elemXml->Value()
+        sdfdbg << "XML Element[" << elemXml->Value()
                << "], child of element[" << _xml->Value()
-               << "] not defined in SDF. Ignoring[" << elemXml->Value() << "]. "
-               << "You may have an incorrect SDF file, or an sdformat version "
-               << "that doesn't support this element.\n";
+               << "], not defined in SDF. Copying[" << elemXml->Value() << "] "
+               << "as children of [" << _xml->Value() << "].\n";
         continue;
       }
     }
+
+    // Copy unknown elements outside the loop so it only happens one time
+    copyChildren(_sdf, _xml, true);
 
     // Check that all required elements have been set
     for (unsigned int descCounter = 0;
@@ -1077,7 +1082,7 @@ static void replace_all(std::string &_str,
 }
 
 /////////////////////////////////////////////////
-void copyChildren(ElementPtr _sdf, TiXmlElement *_xml)
+void copyChildren(ElementPtr _sdf, TiXmlElement *_xml, const bool _onlyUnknown)
 {
   // Iterate over all the child elements
   TiXmlElement *elemXml = nullptr;
@@ -1088,23 +1093,26 @@ void copyChildren(ElementPtr _sdf, TiXmlElement *_xml)
 
     if (_sdf->HasElementDescription(elem_name))
     {
-      sdf::ElementPtr element = _sdf->AddElement(elem_name);
-
-      // FIXME: copy attributes
-      for (TiXmlAttribute *attribute = elemXml->FirstAttribute();
-           attribute; attribute = attribute->Next())
+      if (!_onlyUnknown)
       {
-        element->GetAttribute(attribute->Name())->SetFromString(
-          attribute->ValueStr());
-      }
+        sdf::ElementPtr element = _sdf->AddElement(elem_name);
 
-      // copy value
-      std::string value = elemXml->GetText();
-      if (!value.empty())
-      {
-        element->GetValue()->SetFromString(value);
+        // FIXME: copy attributes
+        for (TiXmlAttribute *attribute = elemXml->FirstAttribute();
+             attribute; attribute = attribute->Next())
+        {
+          element->GetAttribute(attribute->Name())->SetFromString(
+            attribute->ValueStr());
+        }
+
+        // copy value
+        std::string value = elemXml->GetText();
+        if (!value.empty())
+        {
+          element->GetValue()->SetFromString(value);
+        }
+        copyChildren(element, elemXml, _onlyUnknown);
       }
-      copyChildren(element, elemXml);
     }
     else
     {
@@ -1124,7 +1132,7 @@ void copyChildren(ElementPtr _sdf, TiXmlElement *_xml)
           attribute->ValueStr());
       }
 
-      copyChildren(element, elemXml);
+      copyChildren(element, elemXml, _onlyUnknown);
       _sdf->InsertElement(element);
     }
   }
@@ -1314,5 +1322,6 @@ bool erbFile(const std::string &_filename, std::string &_result)
                     std::istreambuf_iterator<char>());
 
   return erbString(data, _result);
+}
 }
 }