+++ Changelog.md
@@ -5,9 +5,18 @@
 1. Removed boost
    * [Pull request #438](https://bitbucket.org/osrf/sdformat/pull-requests/438)
 
+1. SDF DOM: Add copy/move constructors and assignment operator to Geometry
+   * [Pull request 460](https://bitbucket.org/osrf/sdformat/pull-requests/460)
+
+1. SDF DOM: Add copy/move constructors and assignment operator to Material
+   * [Pull request 460](https://bitbucket.org/osrf/sdformat/pull-requests/460)
+
 ## SDFormat 7.0
 
 ### SDFormat 7.0.0 (xxxx-xx-xx)
+
+1. Preserve XML elements that are not part of the SDF specification.
+   * [Pull request 449](https://bitbucket.org/osrf/sdformat/pull-requests/449)
 
 1. Embed SDF specification files directly in libsdformat.so.
    * [Pull request 434](https://bitbucket.org/osrf/sdformat/pull-requests/434)
@@ -17,11 +26,18 @@
 
 1. SDF DOM: Additions to the document object model.
    * [Pull request 433](https://bitbucket.org/osrf/sdformat/pull-requests/433)
+   * [Pull request 441](https://bitbucket.org/osrf/sdformat/pull-requests/441)
+   * [Pull request 442](https://bitbucket.org/osrf/sdformat/pull-requests/442)
    * [Pull request 445](https://bitbucket.org/osrf/sdformat/pull-requests/445)
+   * [Pull request 441](https://bitbucket.org/osrf/sdformat/pull-requests/441)
+   * [Pull request 451](https://bitbucket.org/osrf/sdformat/pull-requests/451)
+   * [Pull request 455](https://bitbucket.org/osrf/sdformat/pull-requests/455)
 
 1. SDF DOM: Additions to the document object model.
-   * [Pull request
-   442](https://bitbucket.org/osrf/sdformat/pull-requests/442)
+   * [Pull request 442](https://bitbucket.org/osrf/sdformat/pull-requests/442)
+
+1. SDF DOM: Add Element() accessor to Gui, JointAxis and World classes.
+   * [Pull request 450](https://bitbucket.org/osrf/sdformat/pull-requests/450)
 
 ## SDFormat 6.0